+++ frontend/app/(main)/(test)/text_extractions/page.tsx
@@ -416,7 +416,8 @@ export default function ExtractionsPage() {
   }
 
   const goBackToResults = () => {
-    router.push('/results')
+    const projectId = activeProject?.id
+    router.push(projectId ? `/projects/${projectId}` : '/projects')
   }
 
   return (
